+++ lumin/nn/models/blocks/head.py
@@ -17,14 +17,14 @@ from ..helpers import CatEmbedder
 from ..initialisations import lookup_normal_init
 from ..layers.activations import lookup_act
 from ..layers.batchnorms import LCBatchNorm1d
-from .gnn_blocks import GravNetLayer
+from .gnn_blocks import GravNetLayer, AbsGraphFeatExtractor, GraphCollapser
 from ....plotting.plot_settings import PlotSettings
 from ....plotting.interpretation import plot_embedding
 from .abs_block import AbsBlock
-from ....utils.misc import to_device
+from ....utils.misc import to_device, is_partially
 from .conv_blocks import Conv1DBlock, Res1DBlock, ResNeXt1DBlock
 
-__all__ = ['CatEmbHead', 'MultiHead', 'InteractionNet', 'RecurrentHead', 'AbsConv1dHead', 'LorentzBoostNet', 'AutoExtractLorentzBoostNet', 'GravNet']
+__all__ = ['CatEmbHead', 'MultiHead', 'GNNHead', 'RecurrentHead', 'AbsConv1dHead', 'LorentzBoostNet', 'AutoExtractLorentzBoostNet']
 
 
 class AbsHead(AbsBlock):
@@ -55,6 +55,8 @@ class AbsMatrixHead(AbsHead):
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init,
                  lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
         super().__init__(cont_feats=cont_feats, cat_embedder=None, lookup_init=lookup_init, freeze=freeze)
+        if row_wise is None: raise ValueError("row_wise is None, possibly a custom inheriting class did not change the class-attribute value. \
+                                               Please set to True or False.")
         self.vecs,self.fpv,self.row_wise,self.lookup_act,self.bn_class = vecs,feats_per_vec,row_wise,lookup_act,bn_class
         self.n_v,self.n_fpv = len(self.vecs),len(self.fpv)
         self._build_lookup()
@@ -345,8 +347,84 @@ class MultiHead(AbsHead):
         return self.flat_head.get_out_size()+self.matrix_head.get_out_size()
 
 
-class InteractionNet(AbsMatrixHead):
+class GNNHead(AbsMatrixHead):
     r'''
+    Encasulating class for applying graph neural-networks to features per vertex.
+    New features are extracted per vertex via a :class:`~lumin.nn.models.blocks.gnn_blocks.AbsGraphFeatExtractor`, and then data is flattened via :class:`~lumin.nn.models.blocks.gnn_blocks.GraphCollapser`
+    
+    Incoming data can either be flat, in which case it is reshaped into a matrix, or be supplied directly into matrix form.
+    Reshaping (row-wise or column-wise) depends on the `row_wise` class attribute of the feature extractor. Data will be automatically converted to row-wise for processing by the grpah collaser.
+
+    .. Note::
+        To allow for the fact that there may be nonexistant features (e.g. z-component of missing energy), `cont_feats` should be a list of all matrix features
+        which really do exist (i.e. are present in input data), and be in the same order as the incoming data. Nonexistant features will be set zero.
+        
+    Arguments:
+        cont_feats: list of all the matrix features which are present in the input data
+        vecs: list of objects, i.e. feature prefixes
+        feats_per_vec: list of features per vertex, i.e. feature suffixes
+        use_int_bn: If true, will apply batch norm to incoming features
+        cat_means: if True, will extend the incoming features per vertex by including the means of all features across all vertices
+        extractor: The :class:`~lumin.nn.models.blocks.gnn_blocks.AbsGraphFeatExtractor` class to instantiate to create new features per vertex
+        collasper: The :class:`~lumin.nn.models.blocks.gnn_blocks.GraphCollapser` class to instantiate to collapse graph to flat data (batch x features)
+        freeze: whether to start with module parameters set to untrainable
+        bn_class: class to use for BatchNorm, default is `nn.BatchNorm1d`
+    '''
+
+    def __init__(self, cont_feats:List[str], vecs:List[str], feats_per_vec:List[str], use_in_bn:bool, cat_means:bool,
+                 extractor:Callable[[Any],AbsGraphFeatExtractor], collapser:Callable[[Any],GraphCollapser], freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
+        super().__init__(cont_feats=cont_feats, vecs=vecs, feats_per_vec=feats_per_vec, freeze=freeze,
+                         row_wise=extractor.func.row_wise if is_partially else extractor.row_wise)
+        self.cat_means,self.use_in_bn = cat_means,use_in_bn
+        if self.use_in_bn: self.bn = LCBatchNorm1d(bn_class(self.n_fpv)) if self.row_wise else bn_class(self.n_fpv)
+        if self.cat_means: self.n_fpv *= 2
+        self.extractor = extractor(n_fpv=self.n_fpv, n_v=self.n_v)
+        sz = self.extractor.get_out_size()
+        self.collapser = collapser(n_fpv=sz[1], n_v=sz[0])
+        self._map_outputs()
+        if self.freeze: self.freeze_layers()
+
+    def forward(self, x:Union[Tensor,Tuple[Tensor,Tensor]]) -> Tensor:
+        r'''
+        Passes input through the GravNet head and returns a flat tensor.
+
+        Arguments:
+            x: If a tuple, the second element is assumed to the be the matrix data. If a flat tensor, will convert the data to a matrix
+        
+        Returns:
+            Resulting tensor
+        '''
+        
+        x = self._process_input(x)  # features per vtx per datapoint
+        if self.use_in_bn: x = self.bn(x)
+        if self.cat_means:
+            if self.row_wise: x = torch.cat([x,x.mean(1).unsqueeze(2).repeat_interleave(repeats=x.shape[1],dim=2).transpose(1,2)],dim=2)
+            else:             x = torch.cat([x,x.mean(2).unsqueeze(1).repeat_interleave(repeats=x.shape[2],dim=1).transpose(1,2)],dim=1)
+        x = self.extractor(x)  # new features per vtx per datapoint
+        if not self.row_wise: x = x.transpose(1,2)
+        x = self.collapser(x)  # features per datapoint
+        return x
+    
+    def _map_outputs(self) -> None:
+        self.feat_map = {}
+        for i, f in enumerate(self.cont_feats): self.feat_map[f] = list(range(self.get_out_size()))
+    
+    def get_out_size(self) -> int:
+        r'''
+        Get size of output
+
+        Returns:
+            Width of output representation
+        '''
+        
+        return self.collapser.get_out_size()
+
+
+class OldInteractionNet(AbsMatrixHead):
+    r'''
+    .. Attention:: This class is depreciated in favour of :class:`~lumin.nn.models.gnn_blocks.InteractionNet`.
+        It is a copy of the old `InteractionNet` class used in lumin<=0.7.2. It will be removed in V0.9
+
     Implementation of the Interaction Graph-Network (https://arxiv.org/abs/1612.00222).
     Shown to be applicable for embedding many 4-momenta in e.g. https://arxiv.org/abs/1908.05318
 
@@ -399,6 +477,8 @@ class InteractionNet(AbsMatrixHead):
         ...                       outfunc_depth=3,outfunc_width=5,outfunc_out_sz=4,agg_method='flatten',
         ...                       do=0.1, bn=True, act='swish', lookup_init=lookup_uniform_init)
     '''
+
+    # XXX remove in V0.9
 
     def __init__(self, cont_feats:List[str], vecs:List[str], feats_per_vec:List[str],
                  intfunc_depth:int, intfunc_width:int, intfunc_out_sz:int,
@@ -1089,7 +1169,6 @@ class GravNet(AbsMatrixHead):
         feats_per_vec: list of features per object, i.e. column headers, feature suffixes
         use_in_bn: whether to add an initial batchnorm layer on the incoming inputs
         cat_means: if True, will extend the incoming features per vertex by including the means of all features across all vertices
-        
         n_s: number of latent-spatial dimensions to compute
         n_lr: number of features to compute per vertex for latent representation
         k: number of neighbours (including self) each vertex should consider when aggregating latent-representation features
@@ -1118,7 +1197,7 @@ class GravNet(AbsMatrixHead):
                  lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
         super().__init__(cont_feats=cont_feats, vecs=vecs, feats_per_vec=feats_per_vec, row_wise=True,
                          lookup_act=lookup_act, lookup_init=lookup_init, freeze=freeze, bn_class=bn_class)
-        store_attr(but=[cont_feats, vecs, feats_per_vec, lookup_act, lookup_init, freeze, agg_methods, bn_class])
+        store_attr(but=['cont_feats', 'vecs', 'feats_per_vec', 'lookup_act', 'lookup_init', 'freeze', 'agg_methods', 'bn_class'])
         self._check_agg_method(agg_methods)
         if not is_listy(self.n_out): self.n_out = [self.n_out]